\name{extended.figures}
\alias{extended.figures}
\title{Generate figures from An Extension of Wilkinson's Algorithm for Position Tick Labels on Axes...}
\usage{extended.figures(samples=100)}
\description{Generate figures from An Extension of Wilkinson's Algorithm for Position Tick Labels on Axes}
\details{Generates Figures 2 and 3 from our paper.}
\value{produces plots as a side effect}
\references{Talbot, J., Lin, S., Hanrahan, P. (2010) An Extension of Wilkinson's Algorithm for Positioning Tick Labels on Axes, InfoVis 2010.}
\author{Justin Talbot \email{jtalbot@stanford.edu}}
\arguments{\item{samples}{number of samples to use (in the paper we used 10000, but that takes awhile to run).}}
