\name{gnuplot}
\alias{gnuplot}
\title{gnuplot's labeling algorithm...}
\usage{gnuplot(dmin, dmax, m)}
\description{gnuplot's labeling algorithm}
\value{vector of axis label locations}
\references{\url{http://www.gnuplot.info/}}
\author{Justin Talbot \email{jtalbot@stanford.edu}}
\arguments{\item{dmin}{minimum of the data range}
\item{dmax}{maximum of the data range}
\item{m}{number of axis labels}}
