\name{heckbert}
\alias{heckbert}
\title{Heckbert's labeling algorithm...}
\usage{heckbert(dmin, dmax, m)}
\description{Heckbert's labeling algorithm}
\value{vector of axis label locations}
\references{Heckbert, P. S. (1990) Nice numbers for graph labels, Graphics Gems I, Academic Press Professional, Inc.}
\author{Justin Talbot \email{jtalbot@stanford.edu}}
\arguments{\item{dmin}{minimum of the data range}
\item{dmax}{maximum of the data range}
\item{m}{number of axis labels}}
