\name{labeling-internal}
\title{Internal labeling objects}
\alias{.coverage}
\alias{.coverage.max}
\alias{.density}
\alias{.density.max}
\alias{.floored.mod}
\alias{.heckbert.nicenum}
\alias{.legibility}
\alias{.simplicity}
\alias{.simplicity.max}
\alias{.wilkinson.nice.scale}
\description{Internal labeling objects.}
\details{These are not to be called by the user.}
\keyword{internal}