\name{matplotlib}
\alias{matplotlib}
\title{Matplotlib's labeling algorithm...}
\usage{matplotlib(dmin, dmax, m)}
\description{Matplotlib's labeling algorithm}
\value{vector of axis label locations}
\references{\url{http://matplotlib.sourceforge.net/}}
\author{Justin Talbot \email{jtalbot@stanford.edu}}
\arguments{\item{dmin}{minimum of the data range}
\item{dmax}{maximum of the data range}
\item{m}{number of axis labels}}
