\name{sparks}
\alias{sparks}
\title{Sparks' labeling algorithm...}
\usage{sparks(dmin, dmax, m)}
\description{Sparks' labeling algorithm}
\value{vector of axis label locations}
\references{Sparks, D. N. (1971) AS 44. Scatter Diagram Plotting, Journal of the Royal Statistical Society. Series C., pp. 327-331.}
\author{Justin Talbot \email{jtalbot@stanford.edu}}
\arguments{\item{dmin}{minimum of the data range}
\item{dmax}{maximum of the data range}
\item{m}{number of axis labels}}
