\name{thayer}
\alias{thayer}
\title{Thayer and Storer's labeling algorithm...}
\usage{thayer(dmin, dmax, m)}
\description{Thayer and Storer's labeling algorithm}
\value{vector of axis label locations}
\references{Thayer, R. P. and Storer, R. F. (1969) AS 21. Scale Selection for Computer Plots, Journal of the Royal Statistical Society. Series C., pp. 206-208.}
\author{Justin Talbot \email{jtalbot@stanford.edu}}
\arguments{\item{dmin}{minimum of the data range}
\item{dmax}{maximum of the data range}
\item{m}{number of axis labels}}
